feat: use new --yes flag on nwui
