+++ .changeset/twenty-books-beg.md
@@ -1,0 +1,5 @@
+---
+'create-expo-stack': patch
+---
+
+Allow for non interactive nwui setup
