+++ cli/__tests__/cli-integration.test.ts
@@ -141,17 +141,14 @@ for (const packageManager of packageManagers) {
 
       expect(fileList).toMatchSnapshot(`${finalFlags.join(', ')}-file-list`);
 
-      // once nwui cli examples are fixed we can remove this
-      if (!finalFlags.includes('--nativewindui')) {
-        const { stderr, stdout, exitCode } = await Bun.$`cd ${projectName} && bun run tsc --noEmit`;
+      const { stderr, stdout, exitCode } = await Bun.$`cd ${projectName} && bun run tsc --noEmit`;
 
-        if (exitCode !== 0) {
-          console.warn('stdout', stdout.toString());
-          console.warn('stderr', stderr.toString());
-        }
-
-        expect(exitCode).toBe(0);
+      if (exitCode !== 0) {
+        console.warn('stdout', stdout.toString());
+        console.warn('stderr', stderr.toString());
       }
+
+      expect(exitCode).toBe(0);
     });
   }
 }
